Fixed sign comparison warning in mxml.

+++ knut/mxml/mxml-string.c
@@ -432,7 +432,7 @@ _mxml_vstrdupf(const char *format,	/* I - Printf-style format string */
 
   bytes = vsnprintf(temp, sizeof(temp), format, ap);
 
-  if (bytes < sizeof(temp))
+  if ((unsigned int)bytes < sizeof(temp))
   {
    /*
     * Hey, the formatted string fits in the tiny buffer, so just dup that...
